requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
lxml>=4.9.0
plotly>=5.3.1
numpy>=1.21.0
python-dateutil>=2.8.2
pytz>=2021.3
pillow>=8.0.0
